refactor: add showcase helper to DRY up the code
