+++ packages/article-topics/article-topics.showcase.js
@@ -2,23 +2,8 @@
 import React from "react";
 import Context from "@times-components/context";
 import { scales } from "@times-components/styleguide";
-import ArticleTopics from "./src/article-topics";
 import topicsData from "./fixtures/topics";
-
-const preventDefaultedAction = decorateAction =>
-  decorateAction([
-    ([e, ...args]) => {
-      e.preventDefault();
-      return ["[SyntheticEvent (storybook prevented default)]", ...args];
-    }
-  ]);
-
-const renderArticleTopics = ({ data, decorateAction }) => (
-  <ArticleTopics
-    onPress={preventDefaultedAction(decorateAction)("onPress")}
-    topics={data}
-  />
-);
+import renderArticleTopics from "./showcase-helper";
 
 export default {
   children: [
