+++ packages/article-topics/article-topics.showcase.web.js
@@ -1,22 +1,5 @@
-/* eslint-disable react/prop-types */
-import React from "react";
-import ArticleTopics from "./src/article-topics";
 import topicsData from "./fixtures/topics";
-
-const preventDefaultedAction = decorateAction =>
-  decorateAction([
-    ([e, ...args]) => {
-      e.preventDefault();
-      return ["[SyntheticEvent (storybook prevented default)]", ...args];
-    }
-  ]);
-
-const renderArticleTopics = ({ data, decorateAction }) => (
-  <ArticleTopics
-    onPress={preventDefaultedAction(decorateAction)("onPress")}
-    topics={data}
-  />
-);
+import renderArticleTopics from "./showcase-helper";
 
 export default {
   children: [
